backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Silent Briefing API"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "clerk-backend-api>=5.0.6",
    "cryptography>=46.0.7",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "lxml>=6.1.0",
    "playwright>=1.58.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "pypdf>=6.10.2",
    "pytest>=9.0.3",
    "python-multipart>=0.0.26",
    "respx>=0.23.1",
    "supabase>=2.28.3",
    "uvicorn[standard]>=0.44.0",
]

[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["briefing"]

[dependency-groups]
dev = [
    "psycopg[binary]>=3.2",
]
